Add some tests for get_jumpdests
