--- tests/src/kakarot/test_execution_context.cairo
@@ -1,25 +1,0 @@
-%lang starknet
-
-from starkware.cairo.common.cairo_builtins import HashBuiltin
-from starkware.cairo.common.alloc import alloc
-from starkware.cairo.common.memcpy import memcpy
-
-from kakarot.model import model
-from kakarot.evm import EVM
-from tests.utils.helpers import TestHelpers
-
-func test__jump{syscall_ptr: felt*, pedersen_ptr: HashBuiltin*, range_check_ptr}() -> model.EVM* {
-    alloc_locals;
-    local bytecode_len: felt;
-    let (bytecode) = alloc();
-    local jumpdest: felt;
-    %{
-        ids.bytecode_len = len(program_input["bytecode"]);
-        segments.write_arg(ids.bytecode, program_input["bytecode"]);
-        ids.jumpdest = program_input["jumpdest"];
-    %}
-    let evm = TestHelpers.init_evm_with_bytecode(bytecode_len, bytecode);
-    let evm = EVM.jump(evm, jumpdest);
-
-    return evm;
-}
